+++ cmd/client.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
-	"github.com/exoscale/egoscale"
 	exov2 "github.com/exoscale/egoscale/v2"
 )
 
@@ -32,7 +31,7 @@ func newCLIRoundTripper(next http.RoundTripper, headers map[string]string) cliRo
 	}
 
 	roundTripper.reqHeaders.Add("User-Agent", fmt.Sprintf("Exoscale-CLI/%s (%s) %s",
-		gVersion, gCommit, egoscale.UserAgent))
+		gVersion, gCommit, exov2.UserAgent))
 
 	for k, v := range headers {
 		roundTripper.reqHeaders.Add(k, v)
@@ -60,27 +59,11 @@ func buildClient() {
 
 	httpClient := &http.Client{Transport: newCLIRoundTripper(http.DefaultTransport, account.CurrentAccount.CustomHeaders)}
 
-	globalstate.EgoscaleClient = egoscale.NewClient(
-		account.CurrentAccount.Endpoint,
-		account.CurrentAccount.Key,
-		account.CurrentAccount.APISecret(),
-		egoscale.WithHTTPClient(httpClient),
-		egoscale.WithoutV2Client())
-
-	// During the Exoscale API V1 -> V2 transition, we need to initialize the
-	// V2 client independently of the V1 client because of HTTP middleware
-	// (http.Transport) clashes.
-	// This can be removed once the only API used is V2.
 	clientExoV2, err := exov2.NewClient(
 		account.CurrentAccount.Key,
 		account.CurrentAccount.APISecret(),
-		exov2.ClientOptWithAPIEndpoint(account.CurrentAccount.Endpoint),
 		exov2.ClientOptWithTimeout(time.Minute*time.Duration(account.CurrentAccount.ClientTimeout)),
-		exov2.ClientOptWithHTTPClient(func() *http.Client {
-			return &http.Client{
-				Transport: newCLIRoundTripper(http.DefaultTransport, account.CurrentAccount.CustomHeaders),
-			}
-		}()),
+		exov2.ClientOptWithHTTPClient(httpClient),
 		exov2.ClientOptCond(func() bool {
 			if v := os.Getenv("EXOSCALE_TRACE"); v != "" {
 				return true
@@ -91,6 +74,6 @@ func buildClient() {
 	if err != nil {
 		panic(fmt.Sprintf("unable to initialize Exoscale API V2 client: %v", err))
 	}
-	globalstate.EgoscaleClient.Client = clientExoV2
+	globalstate.EgoscaleClient = clientExoV2
 
 }