+++ cmd/dns.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
-	"github.com/exoscale/egoscale"
 	exo "github.com/exoscale/egoscale/v2"
 	exoapi "github.com/exoscale/egoscale/v2/api"
 )
@@ -21,8 +20,7 @@ var dnsCmd = &cobra.Command{
 // domainFromIdent returns a DNS domain from identifier (domain name or ID).
 func domainFromIdent(ident string) (*exo.DNSDomain, error) {
 	ctx := exoapi.WithEndpoint(gContext, exoapi.NewReqEndpoint(account.CurrentAccount.Environment, account.CurrentAccount.DefaultZone))
-	_, err := egoscale.ParseUUID(ident)
-	if err == nil {
+	if exo.IsValidUUID(ident) {
 		return globalstate.EgoscaleClient.GetDNSDomain(ctx, account.CurrentAccount.DefaultZone, ident)
 	}
 
@@ -43,8 +41,7 @@ func domainFromIdent(ident string) (*exo.DNSDomain, error) {
 // domainRecordFromIdent returns a DNS record from identifier (record name or ID) and optional type
 func domainRecordFromIdent(domainID, ident string, rType *string) (*exo.DNSDomainRecord, error) {
 	ctx := exoapi.WithEndpoint(gContext, exoapi.NewReqEndpoint(account.CurrentAccount.Environment, account.CurrentAccount.DefaultZone))
-	_, err := egoscale.ParseUUID(ident)
-	if err == nil {
+	if exo.IsValidUUID(ident) {
 		return globalstate.EgoscaleClient.GetDNSDomainRecord(ctx, account.CurrentAccount.DefaultZone, domainID, ident)
 	}