+++ cmd/nlb_update.go
@@ -95,9 +95,4 @@ func init() {
 	cobra.CheckErr(registerCLICommand(nlbCmd, &nlbUpdateCmd{
 		cliCommandSettings: defaultCLICmdSettings(),
 	}))
-
-	// FIXME: remove this someday.
-	cobra.CheckErr(registerCLICommand(deprecatedNLBCmd, &nlbUpdateCmd{
-		cliCommandSettings: defaultCLICmdSettings(),
-	}))
 }