+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## UNRELEASED
 
-## Bug fixes
+### Deprecations
 
-- compute: instance reset default template now falling back to current instance template
+- Remove all the deprecated commands (#526) (Deprecated since v0.51.0)
+
+### Bug fixes
+
+- compute: instance reset default template now falling back to current instance template (#528)
 - compute: remmove uri and tlssni fields when nlb service healthcheck is "tcp"
+- cmd: fix panic if inexistent config file is given (#530)
 
 ## 1.72.0